fix: add craft route

+++ app/business/asset/device/deviceService/deviceServiceImpl/iDeviceTemplateServiceImpl.go
@@ -1,21 +1,25 @@
 package deviceServiceImpl
 
 import (
+	"errors"
 	"github.com/gin-gonic/gin"
 	"nova-factory-server/app/business/asset/device/deviceDao"
 	"nova-factory-server/app/business/asset/device/deviceModels"
 	"nova-factory-server/app/business/asset/device/deviceService"
 	"nova-factory-server/app/utils/baizeContext"
 	"nova-factory-server/app/utils/snowflake"
+	"strconv"
 )
 
 type IDeviceTemplateServiceImpl struct {
-	dao deviceDao.IDeviceTemplateDao
+	dao     deviceDao.IDeviceTemplateDao
+	dataDao deviceDao.ISysModbusDeviceConfigDataDao
 }
 
-func NewDeviceTemplateServiceImpl(dao deviceDao.IDeviceTemplateDao) deviceService.IDeviceTemplateService {
+func NewDeviceTemplateServiceImpl(dao deviceDao.IDeviceTemplateDao, dataDao deviceDao.ISysModbusDeviceConfigDataDao) deviceService.IDeviceTemplateService {
 	return &IDeviceTemplateServiceImpl{
-		dao: dao,
+		dao:     dao,
+		dataDao: dataDao,
 	}
 }
 func (i *IDeviceTemplateServiceImpl) Add(c *gin.Context, template *deviceModels.SysDeviceTemplateSetReq) (*deviceModels.SysDeviceTemplate, error) {
@@ -31,6 +35,25 @@ func (i *IDeviceTemplateServiceImpl) Update(c *gin.Context, template *deviceMode
 	return i.dao.Update(c, data)
 }
 func (i *IDeviceTemplateServiceImpl) Remove(c *gin.Context, ids []string) error {
+	var templateIds []uint64 = make([]uint64, 0)
+	for _, id := range ids {
+		templateId, err := strconv.ParseUint(id, 10, 64)
+		if err != nil {
+			return err
+		}
+		templateIds = append(templateIds, templateId)
+	}
+
+	// 检查协议模板下面是否有数据，有数据不允许删除
+	data, err := i.dataDao.GetByTemplateIds(c, templateIds)
+	if err != nil {
+		return err
+	}
+
+	if data == nil || len(data) > 0 {
+		return errors.New("模板下面有数据，不能重置")
+	}
+
 	return i.dao.Remove(c, ids)
 }
 func (i *IDeviceTemplateServiceImpl) List(c *gin.Context, req *deviceModels.SysDeviceTemplateDQL) (*deviceModels.SysDeviceTemplateListData, error) {
